requirements.txt
tushare~=1.4.21
SQLAlchemy~=2.0.0
pymysql~=1.1.0
pyyaml~=6.0.0
numpy~=1.26.0